se hacen pruebas incluyendo la reducciòn de dimensionalidad

+++ viral/viral_classification/mykameris.py
@@ -20,7 +20,7 @@ def read_fasta(fa_file):
     return data
 
 def cgr(seq, k, order='ACGT'):
-    result = np.zeros(pow(4, k))
+    result = np.zeros(pow(4, k), dtype='uint8')
     #print(result.shape, result)
 
     x = pow(2, k-1)
